build(req): Increase Kinesis shard count for data ingestion
This commit increases the shard count of the AFSFastAPIDataStream Kinesis stream from 1 to 4.

The previous configuration with a single shard was identified as a major performance bottleneck in the data ingestion pipeline, as it limits the throughput to 1,000 records/sec or 1 MB/sec. This change will significantly improve the scalability of the data ingestion pipeline and address the reported performance issues.

This change is made in the `cloudformation/initial_infrastructure.yaml` file. The CloudFormation stack will need to be redeployed for this change to take effect.